Tic Tac Toe game runs as expected.
Additional key code has to been added resulting in a complete functional
Tic Tac Toe game.
The main problem to solve was to get the status bar declaring a win, or
draw and clearing the board.  This was achieved by adding code to the
following:
- updateGame(Player thePlayer, int row, int col) method in class
GameMain.
- hasWon(Player thePlayer, int rowSelected, int colSelected) method of
Board class.
- Drawing of the status bar in paintComponent(Graphics g) method of
GameMain Class.
- isDraw() method in class Board.
- clear() method in class Cells.
Next goals:
- Testing
- Clean up code
- Clearer comments